requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cryptography==42.0.2
dataclasses==0.6
deprecation==2.1.0
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-cron==0.6.0
django-crontab==0.7.1
djangorestframework==3.14.0
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==2.10
iniconfig==2.0.0
openai==1.11.1
packaging==23.2
pluggy==1.4.0
postgrest==0.15.0
postgrest-py==0.4.0
protobuf==4.25.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
realtime==1.0.2
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supabase-py==0.0.2
supafunc==0.3.3
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
websockets==11.0.3
